% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapporteket.R
\docType{package}
\name{rapporteket}
\alias{rapporteket}
\title{rapporteket: A package containing information regarding Rapporteket}
\description{
The rapporteket package provides information regarding Rapporteket: what it
is, how to get stated, regular use and development of content. As such, this
packages cointaines no (or very few) R functions
}
